requirements.txt
transformers>=4.30.0
torch>=1.13.0
diffusers>=0.20.0
pillow>=9.0.0
accelerate>=0.20.0